Instanced objects
Objects are now just their model matrix and a few slots for other
things. Cameras, lights, geometries and materials are attached to an
object. This way, everything is an object. And at the same time we avoid having
10 identical geometries for the 10 objects that are all the same wall. Instead
we declare the geometry once and we associate it to ten objects.