Components and ownership
Added the concept of component. Now camera, light, geometry and material are
contained within the object's component collection. These components are
ultimately located all in a contigious memory block. Which is global and owned
by the componentCollection module. A similar system is used for objects, with
are now all in a contigious memory block, owned by the scene. Creating
components goes through componentCollection and creating objects goes through
scene. The difference being that scene owns them locally to the scene object,
not to the whole program. The logic for this is that objects are only useful
within their scene, while things like definitions of materials or lights can be
shared by objects in the whole program.